+++ protected/humhub/config/__autocomplete.php
@@ -30,7 +30,7 @@ class Yii
  * @property \humhub\components\SettingsManager $settings
  * @property \humhub\modules\notification\components\NotificationManager $notification
  * @property \humhub\components\i18n\Formatter
- * @property \humhub\components\AssetManager $assetManager
+ * @property \humhub\components\assets\AssetManager $assetManager
  * @property \humhub\modules\user\authclient\Collection $authClientCollection
  * @property \yii\queue\Queue $queue
  * @property \humhub\components\Request $request
@@ -39,7 +39,9 @@ class Yii
  * @property \yii\mutex\Mutex $mutex
  * @property \yii\web\User|__WebUser $user
  * @property \yii\caching\ArrayCache $runtimeCache
+ * @property \humhub\components\assets\AssetImageRegistry $img
  * @property \humhub\components\InstallationState $installationState
+ * @property \humhub\components\fs\FilesystemManager $fs
  */
 class __Application
 {