+++ protected/humhub/migrations/m250405_072758_1_18_switch_to_humhub_theme_and_disable_themes.php
@@ -15,6 +15,7 @@ class m250405_072758_1_18_switch_to_humhub_theme_and_disable_themes extends Migr
         $settingsManager = Yii::$app->settings;
 
         // Copy Login Background image
+        /* Migration disabled after v.19
         $oldThemeBasePath = $settingsManager->get('theme');
         foreach (['login-bg.jpg', 'login-bg.png'] as $loginBgFile) {
             $oldFile = $oldThemeBasePath . DIRECTORY_SEPARATOR . 'img' . DIRECTORY_SEPARATOR . $loginBgFile;
@@ -23,6 +24,7 @@ class m250405_072758_1_18_switch_to_humhub_theme_and_disable_themes extends Migr
                 break;
             }
         }
+        */
 
         // Switch to HumHub theme
         $themeAfterMigration = 'HumHub';