+++ protected/humhub/modules/content/components/ContentContainerActiveRecord.php
@@ -9,13 +9,14 @@
 namespace humhub\modules\content\components;
 
 use humhub\components\ActiveRecord;
+use humhub\components\assets\AssetImage;
 use humhub\libs\BasePermission;
-use humhub\libs\ProfileBannerImage;
 use humhub\libs\ProfileImage;
 use humhub\modules\content\models\Content;
 use humhub\modules\content\models\ContentContainer;
 use humhub\modules\content\models\ContentContainerBlockedUsers;
 use humhub\modules\content\models\ContentContainerTagRelation;
+use humhub\modules\space\models\Space;
 use humhub\modules\user\models\User;
 use humhub\modules\user\Module as UserModule;
 use Yii;
@@ -25,9 +26,6 @@ use yii\web\IdentityInterface;
 
 /**
  * ContentContainerActiveRecord for ContentContainer Models e.g. Space or User.
- *
- * Required Methods:
- *      - getProfileImage()
  *
  * @property int $id
  * @property int $visibility
@@ -46,8 +44,9 @@ use yii\web\IdentityInterface;
  * @property-read string $displayName
  * @property-read string|mixed $displayNameSub
  * @property-read ContentContainerModuleManager $moduleManager
- * @property-read ProfileBannerImage $profileBannerImage
- * @property-read ProfileImage $profileImage
+ * @property-read ProfileImage $profileImage @deprecated since v1.19
+ * @property-read AssetImage $image
+ * @property-read AssetImage $bannerImage
  * @property-read string[] $tags
  * @property-read string $wallOut
  *
@@ -66,6 +65,8 @@ abstract class ContentContainerActiveRecord extends ActiveRecord
      */
     private $_moduleManager = null;
 
+    private ?AssetImage $_image = null;
+    private ?AssetImage $_bannerImage = null;
     /**
      * The behavior which will be attached to the base controller.
      *
@@ -91,14 +92,10 @@ abstract class ContentContainerActiveRecord extends ActiveRecord
     public $blockedUsersField;
 
     /**
+     * @deprecated since 1.19
      * @var string
      */
     public $profileImageClass = ProfileImage::class;
-
-    /**
-     * @var string
-     */
-    public $profileBannerImageClass = ProfileBannerImage::class;
 
     /**
      * Returns the display name of content container
@@ -116,25 +113,57 @@ abstract class ContentContainerActiveRecord extends ActiveRecord
      */
     abstract public function getDisplayNameSub(): string;
 
+    public function getImage(): AssetImage
+    {
+        if ($this->_image === null) {
+            $this->_image = new AssetImage([
+                'file' => '/profile_image/' . $this->guid . '.jpg',
+                'defaultOptions' => [
+                    'width' => 150,
+                    'height' => 150,
+                ],
+                'masterOptions' => [
+                    'maxWidth' => 1920,
+                    'maxHeight' => 1080,
+                ],
+                'defaultFile' => ($this instanceof Space)
+                    ? '@webroot-static/img/default_space.jpg'
+                    : '@webroot-static/img/default_user.jpg',
+            ]);
+        }
+        return $this->_image;
+    }
+
+    public function getBannerImage(): AssetImage
+    {
+        if ($this->_bannerImage === null) {
+            $this->_bannerImage = new AssetImage([
+                'file' => '/profile_image/banner/' . $this->guid . '.jpg',
+                'defaultOptions' => [
+                    'width' => 1134,
+                    'height' => 192,
+                ],
+                'masterOptions' => [
+                    'maxWidth' => 1920,
+                    'maxHeight' => 1080,
+                ],
+                'defaultFile' => '@webroot-static/img/default_banner.jpg',
+            ]);
+        }
+        return $this->_bannerImage;
+    }
+
+
     /**
-     * Returns the Profile Image Object for this Content Base
-     *
      * @return ProfileImage
+     * @note use getImage() instead
+     * @deprecated v1.19
      */
     public function getProfileImage()
     {
         return new $this->profileImageClass($this);
     }
 
-    /**
-     * Returns the Profile Banner Image Object for this Content Base
-     *
-     * @return ProfileBannerImage
-     */
-    public function getProfileBannerImage()
-    {
-        return new $this->profileBannerImageClass($this);
-    }
 
     /**
      * Should be overwritten by implementation
@@ -395,7 +424,9 @@ abstract class ContentContainerActiveRecord extends ActiveRecord
      */
     public function getTags(): array
     {
-        $tags = ($this->contentContainerRecord instanceof ContentContainer) && is_string($this->contentContainerRecord->tags_cached)
+        $tags = ($this->contentContainerRecord instanceof ContentContainer) && is_string(
+            $this->contentContainerRecord->tags_cached,
+        )
             ? trim($this->contentContainerRecord->tags_cached)
             : '';
         return $tags === '' ? [] : preg_split('/\s*,\s*/', $tags);
@@ -423,7 +454,7 @@ abstract class ContentContainerActiveRecord extends ActiveRecord
         }
 
         $blockedUsers = $this->getSettings()->get(ContentContainerBlockedUsers::BLOCKED_USERS_SETTING);
-        return empty($blockedUsers) ? [] : explode(',', (string) $blockedUsers);
+        return empty($blockedUsers) ? [] : explode(',', (string)$blockedUsers);
     }
 
     /**