+++ protected/humhub/modules/space/widgets/Image.php
@@ -72,9 +72,7 @@ class Image extends BaseImage
             $imageHtmlOptions['data-bs-title'] = $this->tooltipText ?: Html::encode($this->space->name);
         }
 
-        $isDefaultImage = str_starts_with(basename($this->space->getProfileImage()->getUrl()), 'default_space.jpg');
-
-        if ($isDefaultImage) {
+        if (!$this->space->image->exists()) {
             $imageHtmlOptions['class'] .= ' d-none-space-image'; // Don't replace with `d-none` because it would be removed after changing space
         } else {
             $acronymHtmlOptions['class'] .= ' d-none-space-image'; // Idem