+++ protected/humhub/modules/ui/tests/codeception/unit/ThemeHelperTest.php
@@ -17,10 +17,10 @@ class ThemeHelperTest extends HumHubDbTestCase
 
         // Clear the assets folder
         Yii::$app->assetManager->clear();
-        $this->assertFileNotExists($this->getThemeCssPath($theme));
+        $this->assertFalse(Yii::$app->assetManager->fileExists($this->getThemeCssPath($theme)));
 
         $this->assertTrue(ThemeHelper::buildCss($theme));
-        $this->assertFileExists($this->getThemeCssPath($theme));
+        $this->assertTrue(Yii::$app->assetManager->fileExists($this->getThemeCssPath($theme)));
     }
 
     public function testBuildCssForChildTheme()
@@ -32,7 +32,7 @@ class ThemeHelperTest extends HumHubDbTestCase
     {
         $this->assertInstanceOf(Theme::class, $theme);
         $this->assertTrue(ThemeHelper::buildCss($theme));
-        $this->assertFileExists($this->getThemeCssPath($theme));
+        $this->assertTrue(Yii::$app->assetManager->fileExists($this->getThemeCssPath($theme)));
         $this->assertEquals('#435f6f', $theme->variable('primary'));
     }